fix: prices 테이블 store_id 타입 long으로 변경

+++ backend/src/main/java/com/storerader/server/domain/admin/dto/add/prices/PriceApiItemDTO.java
@@ -8,7 +8,7 @@ public record PriceApiItemDTO(
         @JacksonXmlProperty(localName = "goodInspectDay")
         String inspectDay,
         @JacksonXmlProperty(localName = "entpId")
-        Integer storeId,
+        Long storeId,
         @JacksonXmlProperty(localName = "goodId")
         Integer goodId,
         @JacksonXmlProperty(localName = "goodPrice")
